+++ R/song_param.R
@@ -2,9 +2,9 @@
 #'
 #' \code{song_param} calculates average or extreme values of acoustic parameters of
 #' elements in a song or other level of organization in the signals 
-#' @usage song_param(X = NULL, song_colm = "song",mean_indx = NULL, min_indx = NULL, 
-#' max_indx = NULL, sd = FALSE, parallel = 1, pb = TRUE, na.rm = FALSE, 
-#' weight = NULL)
+#' @usage song_param(X = NULL, song_colm = "song",mean_colm = NULL, min_colm = NULL, 
+#' max_colm = NULL, elm_colm = NULL, elm_fun = NULL, sd = FALSE, parallel = 1, pb = TRUE, 
+#' na.rm = FALSE, weight = NULL)
 #' @param X 'selection_table', 'extended_selection_table' (created 'by.song') or data frame with the following columns: 1) "sound.files": name of the .wav 
 #' files, 2) "selec": number of the selections, 3) "start": start time of selections, 4) "end": 
 #' end time of selections. The ouptut of \code{\link{manualoc}} or \code{\link{autodetec}} can 
@@ -13,11 +13,13 @@
 #' It can be used to label any hierarchical level at which parameters need to be calculated (e.g. syllables, phrases).
 #' Note that 
 #' the function assumes that song labels are not repeated within a sound file.
-#' @param mean_indx Numeric vector with the index of the columns that will be averaged. If \code{NULL} the mean of all numeric columns in 'X' is returned.
-#' @param min_indx Numeric vector with the index of the columns for which the minimum 
+#' @param mean_colm Numeric vector with the index of the columns that will be averaged. If \code{NULL} the mean of all numeric columns in 'X' is returned.
+#' @param min_colm Character vector with the name(s) of the columns for which the minimum 
 #' value is needed. Default is \code{NULL}.
-#' @param max_indx Numeric vector with the index of the columns for which the maximum 
+#' @param max_colm Character vector with the name(s) of the columns for which the maximum 
 #' value is needed. Default is \code{NULL}.
+#' @param elm_colm Character vector with the name(s) of the columns identifying the element labels (i.e. element types). If supplied 'unq.elemts' and 'mean.elemt.count' Default is \code{NULL}.
+#' @param elm_fun Function to be applied to the sequence of elements composing a song. Default is \code{NULL}. Ignored if 'elm_colm' is not supplied. The name of the column containing the function's output is "elm_fun'.  
 #' If  \code{NULL} the mean of all numeric columns in 'X' is returned. 
 #' @param parallel Numeric. Controls whether parallel computing is applied.
 #' It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).
@@ -30,19 +32,21 @@
 #'  for more details.  Default is \code{NULL} (unweighted average).
 #' @return A data frame similar to the input 'X' data frame, but in this case each row corresponds to a single song. The data frame contains the mean
 #'  values for numeric columns for each song. Columns that will be averaged can be defined with
-#'  'mean_indx' (otherwhise all numeric columns are used). Columns can be 
+#'  'mean_colm' (otherwhise all numeric columns are used). Columns can be 
 #'  weighted by other columns in the data set (e.g. duration, frequency range). In addition, the function returns the following song level parameters: 
 #' \itemize{
 #'    \item \code{duration}: length of song (in s)
-#'    \item \code{num.elements}: number of elements (or song units)
+#'    \item \code{num.elmts}: number of elements (or song units)
 #'    \item \code{start}: start time of song (in s)
 #'    \item \code{end}: end time of song (in s)
 #'    \item \code{bottom.freq}: lowest 'bottom.freq' from all song elements (in kHz)
 #'    \item \code{top.freq}: highest 'top.freq' from all song elements (in kHz)
 #'    \item \code{freq.range}: difference between song's 'top.freq' and 'bottom.freq' (in kHz)
 #'    \item \code{note.rate}: number of elements per second
+#'    \item \code{unq.elmts}: number of unique elements (i.e. number element types, only if 'elm_colm' is supplied)
+#'    \item \code{mean.elemt.count}: mean number of times element types are found (only if 'elm_colm' is supplied)
 #'    }
-#'  This function assumes that song labels are not repeated within a sound file.  
+#'    This function assumes that song labels are not repeated within a sound file.  
 #' 
 #' @export
 #' @name song_param
@@ -70,21 +74,22 @@
 #' # caculate song-level parameters for all numeric parameters
 #' song_param(X = sp, song_colm = "song", parallel = 1, pb = TRUE)
 #' 
-#' # caculate song-level parameters selecting parameters with mean_indx
-#' song_param(X = sp, song_colm = "song",mean_indx = 5:10, parallel = 1, pb = TRUE)
+#' # caculate song-level parameters selecting parameters with mean_colm
+#' song_param(X = sp, song_colm = "song",mean_colm = c("dfrange", "duration"), parallel = 1, pb = TRUE)
 #' 
-#' # caculate song-level parameters for selecting parameters with mean_indx, max_indx
-#' # and min_indx and weighted by duration
+#' # caculate song-level parameters for selecting parameters with mean_colm, max_colm
+#' # and min_colm and weighted by duration
 #' song_param(X = sp, weight = "duration", song_colm = "song",
-#' mean_indx = 5:6, min_indx = 7:8, max_indx = 9:10, parallel = 1, pb = TRUE)
+#' mean_colm =  c("dfrange", "duration"), min_colm =  "mindom", max_colm = "maxdom", 
+#'   parallel = 1, pb = TRUE)
 #' 
-#'# with two weights 
-#'song_param(X = sp, weight = c("duration", "dfrange"), song_colm = "song",
-#'mean_indx = 5:9, parallel = 1, pb = TRUE)
+#' # with two weights 
+#' song_param(X = sp, weight = c("duration", "dfrange"), song_colm = "song",
+#' mean_colm = c("kurt", "sp.ent"), parallel = 1, pb = TRUE)
 #'
-#'# with two weights no progress bar
-#'song_param(X = sp, weight = c("duration", "dfrange"), song_colm = "song",
-#'mean_indx = 5:9, parallel = 1, pb = FALSE)
+#' # with two weights no progress bar
+#' song_param(X = sp, weight = c("duration", "dfrange"), song_colm = "song",
+#' mean_colm = c("kurt", "sp.ent"), parallel = 1, pb = FALSE)
 #' }
 #' 
 #' @references {
@@ -93,7 +98,7 @@
 #' @author Marcelo Araya-Salas (\email{araya-salas@@cornell.edu})
 #last modification on may-8-2018 (MAS)
 
-song_param <- function(X = NULL, song_colm = "song", mean_indx = NULL, min_indx = NULL, max_indx = NULL, 
+song_param <- function(X = NULL, song_colm = "song", mean_colm = NULL, min_colm = NULL, max_colm = NULL,  elm_colm = NULL, elm_fun = NULL,
                        sd = FALSE, parallel = 1, pb = TRUE, na.rm = FALSE, weight = NULL)
 {
   # set pb options 
@@ -133,6 +138,19 @@ song_param <- function(X = NULL, song_colm = "song", mean_indx = NULL, min_indx 
   
   if (!any(names(X) == song_colm)) stop("'song_colm' not found")
   
+  
+  # check suplied column names
+  if (!is.null(mean_colm) & !any(names(X) == mean_colm)) stop("at least 1 'mean_colm' supplied not found")
+  
+  if (!is.null(min_colm) & !any(names(X) == min_colm)) stop("at least 1 'min_colm' supplied not found")
+  
+  if (!is.null(max_colm) & !any(names(X) == max_colm)) stop("at least 1 'max_colm' supplied not found")
+  
+  if (!is.null(elm_colm) & !any(names(X) == elm_colm)) stop("'elm_colm' not found")
+  
+  if (!is.null(elm_colm) & !is.null(elm_fun) & if(!is.null(elm_fun)) !is.function(get(as.character(quote(elm_fun)))) else FALSE) stop("'elm_fun' not found")
+  
+  
   if (song_colm == "sound.files") {
     X$song <- X$sound.files
     song_colm <- "song"
@@ -148,10 +166,10 @@ song_param <- function(X = NULL, song_colm = "song", mean_indx = NULL, min_indx 
 
   if (!any(names(X) %in% weight) & !is.null(weight)) stop("'weight' column not found")
   
-  if (!is.null(mean_indx))
-  if (!all(sapply(X[, mean_indx], is.numeric))) stop("not all columns in 'mean_indx' are numeric")
+  if (!is.null(mean_colm))
+  if (!all(sapply(X[, mean_colm], is.numeric))) stop("not all columns in 'mean_colm' are numeric")
   
-  songparam.FUN <- function(Y, song_colm, mean_indx, min_indx, max_indx, weight, na.rm) {
+  songparam.FUN <- function(Y, song_colm, mean_colm, min_colm, max_colm, weight, na.rm) {
     
     Z <- Y[1, , drop = FALSE]
     
@@ -166,35 +184,39 @@ song_param <- function(X = NULL, song_colm = "song", mean_indx = NULL, min_indx 
     wts <- as.vector(unlist(wts))
     } else wts <- rep(1, nrow(Y))
     
-    if (is.null(mean_indx)) {
-      mean_indx <- which(sapply(X, is.numeric))
-      mean_indx <- as.vector(mean_indx[!names(mean_indx) %in% c("selec", "channel", song_colm, "start", "end", "top.freq", "bottom.freq", min_indx, max_indx)])
-      mean_indx <- mean_indx[!is.na(mean_indx)]
+    if (is.null(mean_colm)) {
+      mean_colm <- names(X)[(sapply(X, is.numeric))]
+      mean_colm <- as.vector(mean_colm[!mean_colm %in% c("selec", "channel", song_colm, "start", "end", "top.freq", "bottom.freq", min_colm, max_colm)])
+      mean_colm <- mean_colm[!is.na(mean_colm)]
       }
     
-    for(u in mean_indx) Z[ , u] <- stats::weighted.mean(x = as.vector(c(Y[, u, drop = TRUE])), w = wts, na.rm = na.rm)
+    for(u in mean_colm) Z[ , u] <- stats::weighted.mean(x = as.vector(c(Y[, u, drop = TRUE])), w = wts, na.rm = na.rm)
     
     if (sd){
       W <- Z
       
-      for(u in mean_indx) W[ , u] <- stats::sd(x = as.vector(c(Y[, u, drop = TRUE])), na.rm = na.rm)
+      for(u in mean_colm) W[ , u] <- stats::sd(x = as.vector(c(Y[, u, drop = TRUE])), na.rm = na.rm)
      
       names(W) <- paste0("sd.", names(W))
        }
     
     # minimums
-    if (!is.null(min_indx)) for(u in min_indx)  Z[ ,u]  <- min(x = as.vector(c(Y[, u, drop = TRUE])), na.rm = na.rm)
+    if (!is.null(min_colm)) 
+      for(u in min_colm)  Z[ ,u]  <- min(x = as.vector(c(Y[, u, drop = TRUE])), na.rm = na.rm)
+      
     
     # maximums
-    if (!is.null(max_indx)) for(u in max_indx) Z[ ,u]  <- max(x = as.vector(c(Y[, u, drop = TRUE])), na.rm = na.rm)
+    if (!is.null(max_colm)) for(u in max_colm) Z[ ,u]  <- max(x = as.vector(c(Y[, u, drop = TRUE])), na.rm = na.rm)
     
-    indx <- c(mean_indx, min_indx, max_indx)
+    colm <- c(mean_colm, min_colm, max_colm)
     
-    Z <- Z[, c(names(Z) %in% c("sound.files", song_colm, names(Z)[indx]))]
-    Z <- Z[, match(c("sound.files", song_colm, names(Y)[indx]), names(Z))]
+    #sort columns
+    Z <- Z[, c(names(Z) %in% c("sound.files", song_colm, colm))]
+    Z <- Z[, match(c("sound.files", song_colm, colm), names(Z))]
     
+    # add sd variables  
     if (sd)
-    Z <- cbind(Z, W[, mean_indx, drop = FALSE])
+    Z <- cbind(Z, W[, paste0("sd.", mean_colm), drop = FALSE])
       
     Z$num.elemts <- nrow(Y)
     Z$start <- min(Y$start)
@@ -204,6 +226,28 @@ song_param <- function(X = NULL, song_colm = "song", mean_indx = NULL, min_indx 
     try(Z$freq.range <- Z$top.freq - Z$bottom.freq, silent = TRUE)
     Z$duration <- Z$end - Z$start
     Z$note.rate <- Z$num.elemts/Z$duration
+    
+    # add element parameters
+    if (!is.null(elm_colm))
+    { 
+      Z$unq.elemts <- length(unique(Y[, elm_colm]))
+      Z$mean.elemt.count <- Z$num.elemts / Z$unq.elemts
+    
+      # run element function
+      if (!is.null(elm_fun))  
+      {
+        Z$......temp.name.... <- tapply(Y[, elm_colm], Y[, song_colm], elm_fun)
+        names(Z)[names(Z) == "......temp.name...."] <- as.character(quote(elm_fun))
+      }  
+        }
+
+      
+    # rename columns containing min or max
+    if (!is.null(min_colm)) 
+    names(Z)[names(Z) %in% min_colm] <- paste0("min.", min_colm)
+    
+    if (!is.null(max_colm)) 
+    names(Z)[names(Z) %in% max_colm] <- paste0("max.", min_colm)
     
     return(Z)  
   }
@@ -222,11 +266,20 @@ song_param <- function(X = NULL, song_colm = "song", mean_indx = NULL, min_indx 
   { 
     Y <- X[X$.....SONGX... == i, , drop = FALSE]
     Y <- as.data.frame(Y)
-    return(songparam.FUN(Y, song_colm, mean_indx, min_indx, max_indx, weight, na.rm)) 
+    return(songparam.FUN(Y, song_colm, mean_colm, min_colm, max_colm, weight, na.rm)) 
   }) 
 
   df <- do.call(rbind, out)
   df <- as.data.frame(df)
+  
+  # rename columns
+  rownames(df) <- 1:nrow(df)
+  
+  # add selec label
+  df$selec <- 1
+  
+  # reorder columns
+  df <- sort_colms(df)
   
   return(df)
 }